improvement: re-create position stream on location settings change
It was not possible to change behaviour of the marker (`distanceFilter`)
once the plugin was initialized. For example the marker would move
at speed that was set when the plugin was initialized and if you'd
pass updated `LocationSettings` object, the marker wouldn't react
to new settings. This commit fixes that.

Creation of stream and stream listener were extracted into separate
methods so they can be more easily re-used.